refactor: extract search button and reuse in timetable toolbar
